create parser module, refactor existing methods and objects into it
